requirements.txt
Django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
django-heroku==0.3.1 
gunicorn==19.9.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.0

